Removed race condition in moving files from temporary directory

+++ unpacking.py
@@ -16,6 +16,7 @@ import shutil
 import subprocess
 import sys
 import time
+import tempfile
 
 import peutils
 import requests
@@ -27,7 +28,8 @@ unpack_directory = config.get('clustering', 'unpacking_base_directory')
 static_unpack_directory = unpack_directory + 'static/'
 generic_unpack_directory = unpack_directory + 'generic/'
 
-tmpdir = '/tmp/unpacking/'
+tmpfile_object = tempfile.TemporaryDirectory()
+tmpdir = tmpfile_object.name
 
 packer_sections = {
     '.aspack': 'aspack',
@@ -308,10 +310,7 @@ def unpack_upx(filepath):
     try:
         subprocess.run(["upx", "-d", tmp_path], stdout=subprocess.PIPE, stderr=subprocess.PIPE, check=True)
     except subprocess.CalledProcessError:
-        try:
-            os.remove(tmp_path)             # Delete copy if it could not be unpacked
-        except FileNotFoundError:
-            return []
+        os.remove(tmp_path)                 # Delete copy if it could not be unpacked
         return []                           # Return an empty list if the file could not be unpacked
     except OSError as err:
         print(err)
@@ -367,11 +366,8 @@ def rename_to_sha256(filepath):
         directory = os.path.dirname(filepath)
         sha256sum = hashlib.sha256(rawfile).hexdigest()
         newpath = directory + '/' + sha256sum
-        try:                                # Only rename if it is not already named as the sha256sum
+        if filepath != newpath:             # Only rename if it is not already named as the sha256sum
             shutil.move(filepath, newpath)  # Rename file to the sha256sum
-        except Exception as err:
-            print(err)
-            return None, None
         else:
             return newpath, sha256sum       # Return the new path of the file and the sha256-sum (filename)
     return None, None                       # Return None if the file could not be opened
